Move ManualProxyType
